home screen ready instead of nav bar

+++ feature/navbar-screens/src/main/java/com/example/navbar_screens/home_screen/screen/HomeScreenIntent.kt
@@ -2,5 +2,5 @@ package com.example.navbar_screens.home_screen.screen
 
 sealed class HomeScreenIntent {
     data class UpdateScreenState(val state: HomeScreenState): HomeScreenIntent()
-    data class FetchRandomTitle(val onComplete: () -> Unit): HomeScreenIntent()
+    data class FetchRandomTitle(val onComplete: (Int) -> Unit): HomeScreenIntent()
 }
